requirements.txt
langchain
langchain-openai
langchain-community
python-dotenv
chromadb
streamlit

# For image to recipe project
transformers
torch

# For youtube summarization project
sentence-transformers
yt_dlp
# pip install -q git+https://github.com/openai/whisper.git
# Note: Install FFmpeg (brew install ffmpeg)